// frontend/src/features/actividades/gestion_reparo/models/gestion_reparo.ts
export interface GestionReparo {
  id_gestion_reparo?: number | null;
  id_gestion_rectificatoria?: number | null;
  id_expediente?: number | null;
  id_usuario_solicitante?: number | null;
  subsanar?: boolean | null;
  observaciones?: string | null;
  id_solicitante:number | null;
  id_solicitud:number | null;
  solicitante?: string | null;
  observaciones_reparo?: string | null;
  fecha_ingreso?: string | null;
  
  is_active?: boolean | null;
  row_status?: boolean | null;
  created_by?: number | null;
  created_date?: string | null;
  modified_by?: number | null;
  modified_date?: string | null;
}